moves findAssetBalance helper to fix ciruclar dependency
